+++ hw4/WarClans612/_main.cpp
@@ -1,0 +1,39 @@
+// Developer: Wilbert (wilbert.phen@gmail.com)
+
+#include <pybind11/pybind11.h>
+#include <pybind11/stl.h>
+#include <pybind11/operators.h>
+#include <pybind11/numpy.h>
+
+#include "_matrix.hpp"
+
+namespace py = pybind11;
+
+PYBIND11_MODULE(_matrix, m) {
+    m.def("multiply_naive", &multiply_naive);
+    m.def("multiply_mkl", &multiply_mkl);
+    m.def("multiply_tile", &multiply_tile);
+    py::class_<Matrix>(m, "Matrix", py::buffer_protocol())
+        .def(py::init<size_t, size_t>())
+        //.def(py::init<Matrix const &>())
+        .def(py::init<std::vector<std::vector<double>>&>())
+        .def_property("nrow", &Matrix::nrow, nullptr)
+        .def_property("ncol", &Matrix::ncol, nullptr)
+        .def_buffer([] (Matrix &m) -> py::buffer_info {
+            return py::buffer_info(
+                m.data(),
+                sizeof(double),
+                py::format_descriptor<double>::format(),
+                2,
+                { m.nrow(), m.ncol() },
+                { sizeof(double) * m.ncol(), sizeof(double)}
+            );
+        })
+        .def("__eq__", &Matrix::operator==)
+        .def("__setitem__", [](Matrix &mat1, std::pair<size_t, size_t> i, double v) {
+            mat1(i.first, i.second) = v;
+        })
+        .def("__getitem__", [](Matrix &mat1, std::pair<size_t, size_t> i) {
+            return mat1(i.first, i.second);
+        });
+}
